Github widget

+++ server.js
@@ -6,14 +6,17 @@ global.async = require('async');
 
 /* This is a nasty hack to overwrite globalAuth.json with environment variables
    because atlasboard isn't configurable in any other way */
-var globalAuth;
+var globalAuth = {};
 try {
-  globalAuth = JSON.parse(process.env.GLOBAL_AUTH || '{}');
-  fs.writeFileSync('./globalAuth.json', JSON.stringify(globalAuth));
+  if (process.env.GLOBAL_AUTH || !fs.existsSync('./globalAuth.json')) {
+    globalAuth = JSON.parse(process.env.GLOBAL_AUTH || '{}');
+    fs.writeFileSync('./globalAuth.json', JSON.stringify(globalAuth));
+  }
+  else {
+    globalAuth = JSON.parse(fs.readFileSync('./globalAuth.json'));
+  }
 }
-catch(e) {
-  globalAuth = {};
-}
+catch(e) {}
 
 /* This is a nasty hack to hijack into the express middlewares
    because atlasboard doesn't expose express */
@@ -34,7 +37,9 @@ if (globalAuth.basic && globalAuth.basic.username && globalAuth.basic.password) 
 // Start atlasboard server
 var atlasboard = require('./node_modules/atlasboard/lib/atlasboard');
 atlasboard({ port: process.env.PORT || 5000 }, function (err) {
-  fs.unlink('./globalAuth.json');
+  if (process.env.GLOBAL_AUTH) {
+    fs.unlink('./globalAuth.json');
+  }
   if (err) {
     console.log(err);
   }
